Moved camera out of silly namespace. Added default constructor for colors (black). Fixed compile warnings on visual c++, and added default startup project to cmake. Restructured the way debug draw is rendered to allow for potential tesselation based on distance to camera. In an alternate universe, this CL also ported vectors to use aggregate initializers, but it turns out visual studio very much does not support that

+++ c++/src/TestPlatformAPI/helpers/platform.h
@@ -19,7 +19,7 @@ namespace Platform
     }; };
     
     struct GameConfig {
-        f32 nextFrame;
+        f64 nextFrame;
         RequestFlags::Enum requestFlags;
         bool quit;
     };
@@ -40,7 +40,7 @@ namespace Platform
     struct Input {
         ::Input::Keyboard::State keyboard;
         ::Input::Gamepad::State* pads;
-        s32 padCount;
+        u32 padCount;
     };
     
     struct State {
